Relocate remaining per-estimator configs into helper packages (batch 3+4/4)
Move the final 20 live per-estimator config classes out of config_models.py
into their estimator's utils/<name>_helpers/config.py, re-exported via the
lazy __getattr__ shim so existing imports keep working:

  PANGEO, FSCM, LEXSCM, TASC, SBC, HSC, BVSS, SPCD, NSC, SDID, SparseSC,
  MicroSynth, PPSCM, SequentialSDID, SHC, MLSC, RESCM, SYNDES, SIV, MAREX

Includes the BaseMAREXConfig family (MAREX/LEXSCM/SPCD/SYNDES) and the two
BaseModel-based configs (PANGEO, MLSC). After this, config_models.py holds
only the shared bases (BaseEstimatorConfig, BaseMAREXConfig), the result
contract models, and the relocation shim: ~4700 -> 437 lines. No behavior
change.
