+++ mlsynth/utils/shc_helpers/config.py
@@ -1,0 +1,36 @@
+"""Configuration for the SHC estimator.
+
+Co-located with the helper package; re-exported from
+:mod:`mlsynth.config_models` for backward compatibility.
+"""
+
+from __future__ import annotations
+
+from typing import List, Optional
+from pydantic import Field, model_validator
+from ...exceptions import MlsynthConfigError
+from ...config_models import BaseEstimatorConfig
+
+
+class SHCConfig(BaseEstimatorConfig):
+    m: int = Field(default=1, description="Length of the evaluation window.")
+    bandwidth_grid: Optional[List[float]] = Field(default=None, description="Bandwidth grid for LOOCV.")
+    use_augmented: bool = Field(default=False, description="Use Augmented SHC (ASHC) variant.")
+
+    @model_validator(mode="after")
+    def check_shc_params(self) -> "SHCConfig":
+        if not isinstance(self.use_augmented, bool):
+            raise MlsynthConfigError("'use_augmented' must be a boolean.")
+
+        if self.m <= 0:
+            raise MlsynthConfigError("'m' must be a positive integer.")
+
+        if self.bandwidth_grid is not None:
+            if not self.bandwidth_grid:
+                raise MlsynthConfigError("'bandwidth_grid' cannot be an empty list.")
+            if not all(isinstance(h, (int, float)) for h in self.bandwidth_grid):
+                raise MlsynthConfigError("All elements in 'bandwidth_grid' must be numeric.")
+            if not all(h > 0 for h in self.bandwidth_grid):
+                raise MlsynthConfigError("All bandwidth values must be strictly positive.")
+
+        return self
